rag-system/flask-rag-search/requirements.txt
Flask
faiss-cpu
sentence-transformers
numpy
google-genai
google-cloud-storage
gunicorn
transformers
torch
-e ../common